+++ tests/testthat.R
@@ -1,10 +1,4 @@
 library(testthat)
-library(dplyr)
-library(tidyr)
-library(tibble)
-library(stringr)
-library(purrr)
-library(magrittr)
 
 test_check("dynutils")